refactor: remove UIAlertController extension
Move ping related code inside the View Controller
